[Refactor] #45 도메인 이름 변경
