--- src/main/java/com/umc9th/areumdap/domain/ai/dto/response/AiHistorySummaryDto.java
@@ -1,6 +1,0 @@
-package com.umc9th.areumdap.domain.ai.dto.response;
-
-public record AiHistorySummaryDto(
-    String past,
-    String current
-) {}
