+++ src/main/java/com/umc9th/areumdap/domain/chatbot/dto/response/HistorySummaryResponseDto.java
@@ -1,0 +1,6 @@
+package com.umc9th.areumdap.domain.chatbot.dto.response;
+
+public record HistorySummaryResponseDto(
+    String past,
+    String current
+) {}
